Add files via upload
This is the code that I used for my smart factory project. I used an Arduino for an integrated/automatic control system of a project factory. The main function of this factory was to keep a temperature which in this case I used a thermistor to read the temperature using analog read. The balance set point was at 510 analog read, therefore, if anything was above that value by 10 digits the factory's fan would turn on to cool it down and if the factory got to a temperature of 10 digits below the set point it would turn on an led light to heat up the factory. While all of this is running we also have a water pump taking in water and releasing when it reaches a certain value.